Improve the secret exchange flow
This will make apps trying to concurrently exchange secrets with the auth lobby work.

If the auth lobby doesn't support this new flow, yet, it doesn't matter:
Older versions of the auth lobby will just ignore the message sent and
retrieve the secrets for you anyway (except that this process is then prone to concurrency bugs).