Add username to producer onCompletion
Returning the username is mostly a convenience for developers, but
moving it to the link state might prove useful for multiple accounts at
some point in the future.

Also, a small change in the consuer to reset the username on reseting
the link state. It's not clear yet whether this should also happen in
the producer, and in general whether the producer should keep listening
after a consumer has been linked.