Convert callback interface to event emitter
Instead of expecting callbacks to handle events, the event emitter lets
a developer add event listeners. This interface is more flexible for
developers and for our internal implementation -- developers can
subscribe to events multiple times and we can add events in the future
if we want to.
At the moment, this is incomplete and we need to handle timeouts and
think about cancellation semantics. Various edge cases should also be
considered.