Refactor to remove module global state and more
We want to avoid module global state because it may modules may be
re-instantiated in some cases, which can produce unpleasant and hard to
debug situations. This refactor also removes channel and user
interaction side effects from many of the account linking functions,
which should make them easier to test.
Lastly, the dependency injected channel functions are reduced to a
single createChannel function. It generates send, receive, and close
functions. At the moment, the receive function is unused and maybe won't
be needed.