Add consumer temporary key exchange retry
We can't be certain that users will open a window with the producer before
they request account linking in a consumer. The retry keeps trying until the
consumer receives a session key from a producer.
In addition, retries will be useful in cases where the producer is busy
with another consumer, at least in this version. In a future version,
the producer will queue up consumers that want to be linked.