Add producer linking preflight
We want to check that the producer can actually delegate the account
before broadcast. The producer can delegate if it has the original
keypair the account was created with or if it has a UCAN that delegates
it SUPER_USER capabilities. In the first case, we check that the public
key matches the DID in DNS. In the second case, we check the root issuer
matches the DID in DNS and capabilities are granted in the UCAN.