Add tryParseMessage guard
We want to guard message when they expect JSON with a specific shape,
but instead receive a string (a temporary DID) or a message that doesn't
have the shape they expect. These cases warn because they are likely
noise on the channel.
A few test cases updated and fixed here as well.